rename cmd/benchmark to cmd/loadtest
we're actually doing load testing, not benchmark
