+++ cmd/loadtest/log.go
